refactor(components): refactor dialog (#3369)
* refactor(components): refactor dialog

* refactor: drawer for dialog
