+++ packages/components/dialog/src/dialog.vue
@@ -67,29 +67,28 @@ import { defineComponent, ref } from 'vue'
 import { TrapFocus } from '@element-plus/directives'
 import { Overlay } from '@element-plus/components/overlay'
 
-import {
-  default as useDialog,
-  useDialogProps,
-  useDialogEmits,
-} from './useDialog'
-
-import type { SetupContext } from 'vue'
+import { dialogProps, dialogEmits } from './dialog'
+import { useDialog } from './use-dialog'
 
 export default defineComponent({
   name: 'ElDialog',
   components: {
-    'el-overlay': Overlay,
+    ElOverlay: Overlay,
   },
   directives: {
     TrapFocus,
   },
-  props: useDialogProps,
-  emits: useDialogEmits,
+
+  props: dialogProps,
+  emits: dialogEmits,
+
   setup(props, ctx) {
-    const dialogRef = ref<HTMLElement>(null)
+    const dialogRef = ref<HTMLElement>()
+    const dialog = useDialog(props, ctx, dialogRef)
+
     return {
-      ...useDialog(props, ctx as SetupContext, dialogRef),
       dialogRef,
+      ...dialog,
     }
   },
 })
